Rename: Refactor WeatherForecastModel in calendar screen
